withTimeSeriesTooltip HOC
